Sprint 4 Completed

+++ src/main/java/com/codependentvariables/aiandme/controller/ProfileController.java
@@ -1,10 +1,9 @@
 package com.codependentvariables.aiandme.controller;
 
+import com.codependentvariables.aiandme.Icon;
 import com.codependentvariables.aiandme.model.User;
-import com.codependentvariables.aiandme.navigation.Router;
-import com.codependentvariables.aiandme.navigation.Toast;
-import com.codependentvariables.aiandme.navigation.ToastMessageType;
-import com.codependentvariables.aiandme.navigation.View;
+import com.codependentvariables.aiandme.modules.*;
+import com.codependentvariables.aiandme.services.UserDataDeletionService;
 import com.codependentvariables.aiandme.services.UserService;
 import com.codependentvariables.aiandme.state.AppState;
 import com.codependentvariables.aiandme.validation.ValidationEntry;
@@ -16,12 +15,14 @@ import javafx.fxml.FXML;
 import javafx.scene.control.Button;
 import javafx.scene.control.CheckBox;
 import javafx.scene.control.TextField;
+import javafx.scene.shape.SVGPath;
 
 import java.util.Objects;
 
 public class ProfileController {
     private final AppState appState = AppState.getInstance();
     private final UserService userService = UserService.getInstance();
+    private final UserDataDeletionService dataDeletionService = UserDataDeletionService.getInstance();
 
     private User currentUser;
 
@@ -32,9 +33,17 @@ public class ProfileController {
     @FXML
     public Button cancelNameButton;
     @FXML
+    public SVGPath editNameSvg;
+    @FXML
+    public SVGPath cancelNameSvg;
+    @FXML
     public TextField emailField;
     @FXML
     public Button emailButton;
+    @FXML
+    public SVGPath editEmailSvg;
+    @FXML
+    public SVGPath cancelEmailSvg;
     @FXML
     public Button cancelEmailButton;
     @FXML
@@ -65,18 +74,21 @@ public class ProfileController {
         }
 
         nameField.setText(currentUser.getName());
-        nameButton.textProperty().bind(
+        editNameSvg.contentProperty().bind(
                 Bindings.when(nameField.disableProperty())
-                        .then("E")
-                        .otherwise("S")
+                        .then(Icon.PENCIL)
+                        .otherwise(Icon.PENCIL) // TODO: change to save when icon is added
         );
+        cancelNameSvg.setContent(Icon.CANCEL);
         cancelNameButton.visibleProperty().bind(nameField.disableProperty().not());
+
         emailField.setText(currentUser.getEmail());
-        emailButton.textProperty().bind(
+        editEmailSvg.contentProperty().bind(
                 Bindings.when(emailField.disableProperty())
-                        .then("E")
-                        .otherwise("S")
+                        .then(Icon.PENCIL)
+                        .otherwise(Icon.PENCIL) // TODO: change to save when icon is added
         );
+        cancelEmailSvg .setContent(Icon.CANCEL);
         cancelEmailButton.visibleProperty().bind(emailField.disableProperty().not());
 
         boolean hasMfa = currentUser.getTotpSecret() != null;
@@ -172,7 +184,10 @@ public class ProfileController {
     }
 
     @FXML
-    private void viewCheckIns() { Router.navigateLayout(View.CHECKIN_HISTORY); }
+    private void navigateCheckInHistory() {
+        Router.navigateLayout(View.CHECK_IN_HISTORY);
+    }
+
     @FXML
     private void exportData() {
         throw new RuntimeException("Export user data not implemented.");
@@ -180,13 +195,32 @@ public class ProfileController {
 
     @FXML
     private void clearData() {
-        throw new RuntimeException("Clear user data not implemented.");
+        dataDeletionService.deleteCurrentUserData();
+        Toast.addMessage("Data Cleared", "All data associated with this account has been deleted.", ToastMessageType.INFORMATION);
+    }
+
+    @FXML
+    private void clearCheckIns() {
+        dataDeletionService.deleteCurrentUserCheckIns();
+        Toast.addMessage("Check-ins Deleted", "All check-ins associated with this account have been deleted.", ToastMessageType.INFORMATION);
+    }
+
+    @FXML
+    private void clearQuizAttempts() {
+        dataDeletionService.deleteCurrentUserQuizAttempts();
+        Toast.addMessage("Quiz Attempts Deleted", "All quiz attempts associated with this account have been deleted.", ToastMessageType.INFORMATION);
     }
 
     @FXML
     private void deleteUser() {
-        userService.deleteCurrentUser();
-        Router.navigateLayout(View.HOME);
-        Toast.addMessage("Account Deleted", "We'll miss you!", ToastMessageType.INFORMATION);
+        Dialogue.confirmationWithCancel(result -> {
+            if (result == null || !result) {
+                return;
+            }
+
+            userService.deleteCurrentUser();
+            Router.navigateLayout(View.HOME);
+            Toast.addMessage("Account Deleted", "We'll miss you!", ToastMessageType.INFORMATION);
+        });
     }
 }
